create register decorator to register commands
